requirements.txt
psycopg[binary]
python-telegram-bot
requests
